refactor: spectator fetcher 의존성 분리
